clangd/clang says ...
controllers/sensors/impl/software_knock.cpp|98 col 33-50 warning| Result of integer division used in a floating point context; possible loss of precision

+++ firmware/controllers/sensors/impl/software_knock.cpp
@@ -95,7 +95,7 @@ void initSoftwareKnock() {
 
 		efiPrintf("Knock sense configuring filter with frequency %.2f khz", freqKhz);
 
-		knockFilter.configureBandpass(KNOCK_SAMPLE_RATE, 1000 * freqKhz, 3);
+		knockFilter.configureBandpass((int)KNOCK_SAMPLE_RATE, 1000 * freqKhz, 3);
 
 		efiSetPadMode("knock ch1", KNOCK_PIN_CH1, PAL_MODE_INPUT_ANALOG);
 #if KNOCK_HAS_CH2
